Revert "Osc organization"
