+++ src/constants/oscillator.js
@@ -1,13 +1,3 @@
-export const oscCfg = {
-  // number of oscillators
-  osc: 2,
-  // number of voices per osc
-  voices: 5,
-  // octave variation from the main octave
-  octave: 0,
-  // osc type
-  type: 'sine',
-  // max amp in relation to the other osc
-  amplitude: .5
-};
+export default {
 
+};
